try to generalize integrator
